+++ common/src/python/datastore/forms_store.py
@@ -6,7 +6,7 @@ from json import JSONDecodeError
 from typing import Dict, List, Literal, Optional
 
 from flywheel_adaptor.flywheel_proxy import ProjectAdaptor
-from keys.keys import DefaultValues
+from keys.keys import DefaultValues, MetadataKeys
 
 log = logging.getLogger(__name__)
 
@@ -108,7 +108,7 @@ class FormsStore():
         title = ('Visits for '
                  f'{project.group}/{project.label}/{subject_lbl}/{module}')
 
-        search_col = f'{DefaultValues.FORM_METADATA_PATH}.{search_col}'
+        search_col = f'{MetadataKeys.FORM_METADATA_PATH}.{search_col}'
         columns = [
             'file.name', 'file.file_id', "file.parents.acquisition",
             "file.parents.session", search_col
@@ -116,7 +116,7 @@ class FormsStore():
 
         if extra_columns:
             for extra_lbl in extra_columns:
-                extra_col = f'{DefaultValues.FORM_METADATA_PATH}.{extra_lbl}'
+                extra_col = f'{MetadataKeys.FORM_METADATA_PATH}.{extra_lbl}'
                 columns.append(extra_col)
 
         filters = f'acquisition.label={module}'